More informative error message when a db fails
When the DB fails, this will show the user a "Could not update map: Database Unavailable" as opposed to "Could not update map: Error".

Basing this on a 503 (Service Unavailable) vs. the 500 (Internal Server Error) is probably more appropriate as well.

Also fix the tab before the cleanupDb() to be 4 spaces.